__viejo/__viejo_discord/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
discord.py==2.3.0
fonttools==4.40.0
frozenlist==1.3.3
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
multidict==6.0.4
numpy==1.25.0
packaging==23.1
Pillow==9.5.0
pyparsing==3.1.0
python-dateutil==2.8.2
six==1.16.0
yarl==1.9.2
